Internal cleanup: create xarray.core.duck_array_ops
