+++ xarray/testing.py
@@ -5,7 +5,7 @@ from __future__ import print_function
 
 import numpy as np
 
-from xarray.core import ops
+from xarray.core import duck_array_ops
 
 
 def _decode_string_data(data):
@@ -21,9 +21,10 @@ def _data_allclose_or_equiv(arr1, arr2, rtol=1e-05, atol=1e-08,
         arr2 = _decode_string_data(arr2)
     exact_dtypes = ['M', 'm', 'O', 'U']
     if any(arr.dtype.kind in exact_dtypes for arr in [arr1, arr2]):
-        return ops.array_equiv(arr1, arr2)
+        return duck_array_ops.array_equiv(arr1, arr2)
     else:
-        return ops.allclose_or_equiv(arr1, arr2, rtol=rtol, atol=atol)
+        return duck_array_ops.allclose_or_equiv(
+            arr1, arr2, rtol=rtol, atol=atol)
 
 
 def assert_equal(a, b):
